requirements.txt
asgiref==3.8.1
Django==5.0.6
django-js-asset==2.2.0
django-mptt==0.16.0
djangorestframework==3.15.1
sqlparse==0.5.0
tzdata==2024.1
